Fixes is_accessed_by. CNE_01700 is not working.
The BCE has been updated to include a group of items that should be
tagged by the CNE_01700 rule. However, it's not happening. I don't see
anything wrong with the way the CNE_01700 rule is written and a quick
overview of the generated model would lead me to conclude that it indeed
should tag something. Likely there are still issues with the way
properties are turned into Kodkod formulas.