Limits the creation of atom-relations to "String".
The need for those relations is limited to the access of their atom in
the formula. As, with the exception of String elements, the formula
cannot directly reference an atom, only String atoms require their own
relation. Going even further, only the String element referenced in the
formula need their own relation.